Fixes seg fault when clicking on a non-selected goal marker
When clicking, the feedback callback was called. Inside the callback, the marker was searched in the list and set to selected. This called the selectionChanged callback, where the marker gets replaced by a new one.
This is all done before the feedback callback finishes, which means the old marker signal is deleted before the callback is completed, thus leading to a seg fault. Now the selection is scheduled as a MainLoopJob, when we are sure the marker feedback callback has finished