disable path display when done animating
After animating the plan, turns off the plan robot display so it does not
clutter up the screen.  If trails are enabled the trails will still be
displayed.  Also fixes trails so that while animating only trails before the
current position are visible (without this the animation is not visible as it
is obscured by the trails).

Fixes issue #151