update pose of all markers when any marker moved
Having several end-effector markers attached to a group (e.g. a multi-
fingered hand having an end-effector per fingertip and an end-effector
for the hand base), all markers need to update their pose on any motion
of any marker. In the example: if the hand base is moved, the fingertip
markers should be moved too.