Remove of collision Object when attaching it with apply planning scene service
This is part of the issue #103.

The current state is that the collision object will be removed from the planning scene when attaching
it to the robot with the asynchronous attachObject function of the move_group_interface.

So we want the same behavior with the synchronous apply planning scene service.

To reach this we need to call the processAttachedCollisionObjectMsg, the same function as in attachObject,
to attach the object to the robot and not in the robotStateMsgsToRobotState, because we can't get access to
the collision objects in the robot state. Therefore we give this method a scene without the attached objects
and attach them with the process funcion afterwards.

In this effect we add a error message for the case that the roboState is_diff is false and someone uses
a different operation than ADD.