msa: fix highlighting when clicking on cells
I somehow missed this while testing.
For some reason Qt sends currentCellChanged with the *previous*
cell's indices when a new cell is clicked.

Thus, I've changed the code to highlight links on *both* events.
This works, because cellClicked of the new cell is handled after
currentCellChanged of the previously-selected cell.