remove default grasp from pick-pipeline
If a user uses the pick(std::string) interface that means the code
will call the external planning service. If this service is not available
or does not provide grasps, this block added a default grasp.

This is quite dangerous actually, because when the service is not available
this probably means that it hasn't been started or died. In this case the user
didn't get an error message but instead got a more or less random grasp
that got processed instead and that could induce unwanted movements of the manipulator.

Yes this changes semantics in a stable release, but I consider it to be a safety issue.