wait a second before updating "current" in RViz (#291)
This is a poor-man's-replacement for rhaschke's work that
waits for the current robot state. This can be removed after
his work is merged. (See https://github.com/ros-planning/moveit/pull/232)

Without the additional sleep the automatic update of the start state
will pick a point near the end of the executed trajectory instead of
the current state. Let's give the monitor a moment to catch up.