remove grasp service support from pick_place's fillGrasp
The combination of the service support here and an optional hard-coded default grasp
is problematic because if the service is not available for any reason, but should be called,
the pick-request will end up with the hard-coded grasp.
This can create unexpected trajectories.

With MoveGroupInterface::planGraspsAndPick, a replacement for the removed functionality
is available.

This removes the dependencies on manipulation_msgs and household_objects_database_msgs.
(These were not even consistently specified in package.xml and CMakeLists.txt)