ompl_interface: uniform & simplified handling of the default planner
PR #340 introduced additional logic to select the configuration "group_name[default]" as
the default configuration and sets it to RRTConnect.
But there was already logic in place for this!
The previous default configuration is plainly called "group_name".

This basically reverts #340 and sets the (old and new) default config to RRTConnect.

Additionally, this unifies the default planner setup within the ompl_interface
with the variable "default_planner_config" that can be specified for
each group. This was previously only implemented in the RViz plugin
and could be retrieved via `MoveGroupInterface::getDefaultPlannerId`,
but is not set there by default.
As a result the planner specified in the variable was *always* selected in RViz,
but *never* when an innocent user requested a motionplan in their code,
without explicitly setting the default planner.