update link transforms in UnionConstraintSampler::project
This extends #186 (0119d584bd77a754ed5108d0b222cbcb76326863).

Scott Paulin found that the UnionConstraintSampler should explicitly
update the LinkTransforms of the returned samples, because one of the
other samplers could require them to be valid.

Back then, we missed that the sampler also provides a `project` method
that has the same limitation. So I update the transforms there too.