re-enable support for cmake 2.8.11
842b590cbe3cafa1e9e6e00989235ef26a6fcf91 introduced a minimum cmake of 2.8.12.
This was fully reasonable and worked at that time.
But the obsolete ubuntu saucy only provided 2.8.11 and it seems OSRF
now forces us to support this version more-or-less "for technical reasons".

This commit moves back to 2.8.11 and follows the advice of qt developers
to circumvent an old compile-flags problem with Qt5.