PlanningScene: Don't reset stored object colors with each new message
The stored colors were cleared whenever a new non-empty color entry
was received. This is fine as long as one always sends the colors of all
scene objects *together*, but breaks as soon as someone adds objects
incrementally and only provides the color of the newly added object.

In order to prevent dangling color information, I explicitly remove
the relevant color entries after the respective object is removed.