Handle multiple shapes in an attached collision object
Objects with multiple collision shapes leave the question what the object's
internal frame is, because a pose is specified *for each shape* individually.

The RobotState::getFrameTransform function handled this by returning the first
registered pose as frame of reference.
This is a perfectly valid approach, so I changed the warning there to a debug message.

RobotState::knowsFrameTransform, however, returned false if asked whether the state knows the
transform for the object. That's inconsistant behavior: if `getFrameTransform` returns a valid result,
then the indicator should also reflect that.

Among other things, the pick-pipeline fails with objects with multiple collision shapes
without this change. I consider this a bug-fix, so I request to merge this into indigo (and forward-port it).