improve RobotState display [kinetic] (#465)
* robot state display: fix segfault on topic change

loadRobotModel should be run in the update thread.
Calling it in the wrong thread leads to mysterious segfaults over here.

* robot state display: subscribe on enable / unsubscribe on disable

The display didn't connect in onEnable.
It only did after the user changed the topic name.

Also, disconnect onDisable. There's no need to keep listening to the topic
until the display is enabled again.