New panel with a slider to control the visualized trajectory (#491)
* New panel with a slider controlling the planned path

This panel is integrated in the trajectory visualization, that means for the Trajectory
display and the MotionPlanning display appears a new panel.

This is an extension to the displayed planned path, currently the planned path is
shown once after it is planned, if the loop mode is enabled the animation will
be repeated. These behaviors will be kept.

This panel includes a slider which controls this animated path, it is possible now
to display the single waypoints of the trajectory. There is also a play/pause button
which starts the animation from the current waypoint to the end and to pause it again.
Pressing the play button after the animation is finished replays the trajectory.
This can also be used to pause the loop mode.

* Simplified logic of the animated planned path with the slider

the changed logic clarify the seperate cases we have with the loop mode and the slider.