robot display: Don't reload robot model upon topic change
Instead just reset it to the default configuration.
This avoids needless urdf-parsing and retains a valid kstate_.
For changes to the robot description parameter,
there is a separate callback function.

On Ubuntu 14.04 indigo this also fixes a segfault during startup,
that seems to be related to calling loadRobotModel() twice in a row.

Fixes #528

Conflicts:
	moveit_ros/visualization/robot_state_rviz_plugin/src/robot_state_display.cpp