ikfast_kinematics_plugin: Write XML files as UTF-8 (#514)
* ikfast_kinematics_plugin: Write XML files as UTF-8

No more garbled author names for people with an umlaut in their name! :)

* ikfast_kinematics: Don't modify package.xml twice

Removes some duplicated code

* ikfast_kinematics_plugin: Always regenerate package.xml

This fixes the following bug: If the current package.xml was encoded as ASCII instead of UTF-8, but neither the build nor run depends nor the exports are changed, the package.xml wasn't regenerated to switch to UTF-8.