Adapt pick pipeline to function without object (#599)
Selective Backport / Did not add default parameter for `planGraspsAndPick`

* Use planGraspsAndPick with empty string

Make it possible to use planGraspsAndPick even though the user did not provide
any object.
Now, if planGraspsAndPick is called with an empty string then planGraspsAndPick is called again
with an empty moveit_msgs::CollisionObject.

This change causes trouble in the executeAttachObject function, because there is no object to be attached.
The function returns false and the pick is being abborted after the grasp.
Thus we just return true in that function when the id of the object is an empty string.