Fix MotionPlanning Display auto active bug
The MotionPlanning display got activated every time rviz was started,
even though it was disabled in the configuration.
This behaviour was caused by a function that enabled the display whenever
the MotionPlanning panel was added in rviz.

To fix this issue the function and its calls were removed.
The display's state behaves according to the configuration on startup,
but the display does not get enabled anymore when adding the panel.