make operator bool() explicit in MoveItErrorCode
this protects from bugs in code comparing one the anonymous enums defined in moveit_msgs::MoveItErrrorCodes to the struct. Previously comparing

MoveItErrorCode::PLANNING_FAILED == MoveItErrorCode(MoveItErrorCode::PLANNING_FAILED)

would always return "false" except for MoveItErrorCode::SUCCESS. The reason is that the class MoveItErrorCode was implicitly downcast (via operator bool())
to a bool and then compared to the enum value which was also downcast first to int and then to bool.

This has led to some hard to catch bugs in code like the following in our software:

MoveItErrorCode ret = plan(...);
if (ret == MoveItErrorCode::NO_IK_SOLUTION) { // always false
    ROS_ERROR("No ik solution found but you'll never know");
}

I think preserving some of the type safety is worth adding a bit of explicitness in the two other files that needed changes.