Adds parameter lookup function for kinematics plugins
Some kinematics plugins rely on certain parameters to be set in the private namespace of their node.
Running multiple instances requires each node to be initialized with these parameters which can be
very inconvenient in some setups.
This commit resolves the issue by enabling plugins to lookup parameters from the
namespace 'robot_description_kinematics' so that they can be defined in the kinematics.yaml.