Change order of parameter lookup
Parameters are queried in order of the specified group hierarchy.
That is parameters are first searched in the private namespace
then in the subroup namespace and finally in the group namespace.
This order maintains default behavior by keeping the private namespace
as the predominant configuration but also allows groupwise specifications.