Added warning when the first two points of a trajectory are identical.
This change came about while trying to run the Moveit Tutorials on the moveit website using the Python API. The website appends the starting point to the waypoints[] list, which results in errors because the path planner will try and append a waypoint to the path that does not monotonically increase with time.

I have since added this warning to alert the user of the plausible cause of an error as well as updated the tutorial to reflect a less error prone example.