Use CollisionObject id if ObjectColor id is empty
When calling addCollisionObjects or applyCollisionObjects
with colors for the objects, each ObjectColor id had to be set to
the corresponding object id.
This can lead to confusion as each method is already given a vector
of CollisionObjects, so that one might think each ObjectColor is
applied to the CollisionObject with the corresponding index.

Now, in case the ObjectColor id field is left empty, the id of the
CollisionObject with the corresponding index is used instead.
This is done until one of the ObjectColors in the vector contains
an id.

If there are still empty ObjectColor ids left after an ObjectColor
already contained a non-empty id an error message is printed.