Added check for whitespace in planning group name
When the planning group name has any unintentional whitespace at the start or end, the group name is accepted. However, if we try adding robot poses, it will show the following error message. In this case, the planning group name was manipulator with an added whitespace at the end. 

> Unable to find joint model group for group: manipulator  Are you sure this group has associated joints/links?

I have added a check for any additional whitespace, which reminds the user to not use whitespace in the group name. This similar issue had been raised in a [ROS Answers question](https://answers.ros.org/question/290983/error-in-planning-group-with-moveit-setup-assistant/) as well as [this PR](https://github.com/ros-industrial/industrial_training/pull/199) related to the ROS Industrial training material.