allow execution of empty trajectories
This is useful because the plan_execution allows arbitrary callbacks
to run after successful execution of individual plan_components_.

So although the *trajectory* might be empty we miss to run the
associated callbacks if we just skip the component/trajectory altogether.

This mechanism is very convenient for modular plans where some components
only invoke side effects.