chomp tests: add moveit includes before
This resolves problems when a system version of moveit is overlayed.
It could still trigger new bugs when `moveit_ros_planning_interface`
is installed in a workspace other than the current one where chomp_interface
is compiled, but this is not a big issue here because both packages
are part of the same repository and are usually compiled together.