Allow AttachedBody frames as eef_links
If a motion plan request is sent through the move group interface, and its end_effector_link is a named frame on an AttachedBody instance, the function added in this commit makes the motion plan request valid by transforming the eef_link to a robot link. It uses a MoveGroupCapability because the MoveGroupInterface only tracks the robot joints, not the attached bodies. Doing it via a request to the PlanningScene would be possible as well, but this is the first prototype.

"validateConstraints" is not a good name for a function that forces the constraints to become valid, but I couldn't come up with anything better on the spot.