add support for fcl 0.6
fcl 0.6 changed paths to relevant headers as well as changed the
name of many types compared to fcl 0.5.
This patch adds the header fcl_compat.h to check for the fcl
version and include fcl headers accordingly.
The old types are exchanged to the newer ones throughout the
package.
Compability to fcl 0.5 is ensured by creating type aliases that
map the old types to the new ones.