simplify choosing start and goal state
The update buttons were removed and the goal and start state are now
updated directly after an option was clicked in the combo box.
Random states can be updated by clicking on that choice repeatedly.

The combo boxes were previously in a QToolBox where only one of the
selections was visible at a time.
Because the removal of the update buttons saves a lot of space the
QToolBox was removed as well, so that both combo boxes are visible at
all times.