Provide a tool to generate arbitrary constraint db
Previously, we only shipped a demo_construct_state_database.cpp
that hard-coded all parameters for a specific setup.

This work generalizes the code and provides a utility that
can be configured via ROS parameters.

Assuming a MoveIt configuration is running,
an example launch file to use the utility might look like this:

<launch>
  <node name="generate_state_database" pkg="moveit_planners_ompl" type="generate_state_database" output="screen">
    <param name="use_current_scene" value="false"/>
    <param name="planning_group" value="arm"/>
    <remap from="~output_folder" to="move_group/constraint_approximations_path"/>

<rosparam ns="constraints">
path_constraint:
  name: tool0_upright
  constraints:
  - type: orientation
    frame_id: world
    link_name: tool0
    orientation: [0, 0, 0]
    tolerances: [0.01, 0.01, 3.15]
    weight: 1.0
</rosparam>

  </node>
</launch>