Make FCL shape cache thread-local
Currently there are two shared caches of FCLGeometry objects, one for world objects
and one for attached objects. If a Shape is not found in the matching one, it is either
moved from the other one or created (and stored) if it does not exist. For an existing
entry the shared_ptr is checked to be unique, if it isn't a new Geometry is created from
scratch as well.

This all works quite well in a single threaded world but if there are many threads cache
entries will basically always be in use by some other thread.

This commit replaces the global caches with thread-local ones. All locking can then be
dropped as well. A new cache will be created per thread automatically and is freed once
the threads finishes.