Add support for multi-query planning in OMPL (#1799)
If enabled, the internal state of of planners does not get cleared.
This can be enabled per planner. It only makes sense for roadmap-based
planners (or if the start state does not change). Additionally, for
PRM, PRM*, LazyPRM, and LazyPRM* it is possible (when using OMPL 1.5.0)
to load/save roadmaps from disk upon creation/destruction. This is
set in ompl_planning.yaml like so:

planner_configs:
  PersistentPRM:
    type: geometric::PRM
    multi_query_planning_enabled: true
    store_planner_data: true
    load_planner_data: true
    planner_data_path: /tmp/roadmap.graph

This commit also fixes a bug for PRM-family planners where they would
always run as optimizing planners and not terminate when the first
solution was found. This occurs when the default optimization objective
is set when none are specified. The fix is to not set an objective
when none are specified.

Co-Authored-By: Robert Haschke <rhaschke@users.noreply.github.com>