added options to start and end state dropdowns
The current default behavior is unintuitive, and may be wrong. Changing
the start state in the graphical interface does not impact the plan
generated--it just starts from the robot's current state. The sematics
of "<current>" are unclear, and apparently different for the start state
and the end state.

I added the "Query Start State" option to the Start State dropdown menu,
and similarly "Query Goal State" for the Goal State menu. These options
intuitively correspond to the states defined in the graphical interface.
I also made them the default options.