remove PSI from rviz display
The display maintains a PSM listening to move_group's PSM.
Using an additional PSI (connecting to the move_group via services) is bad design
and the use of it - to poll via ros-services - is even worse.

Implemented the same functionality through the local PSM.

Note that the second argument of updateDetectedObjectsList was never used.