MP display: change internal shortcut Ctrl+R to Ctrl+I
Ctrl+R is used by RViz to rename displays.
Overlaying the sequence semi-breaks the behavior
and the rename button gets selected, but not pressed.

This is an internal shortcut used in case the interactive markers
of the display fail to update correctly. This is seldomly (never?)
used, so remapping does not hurt anyone.

To the best of my knowledge Ctrl+I is not used in standard RViz.