osqp installed from trajopt_utils instead of trajopt_sco
osqp_FOUND is false if osqp is pulled in trajopt_sco. osqp can not be found from the same CMakeLists that was used to pull down its repo and install it in build directory. A workaround is to install osqp from trajopt_utils and check if it exists in trajopt_sco. trajopt_sco is dependent on trajopt_utils

Another fix:  trajopt_problem_ from trajopt_interface needed to be corrected as a shared_pointer.