Update noetic.rosinstall
Only include missing upstream packages to be built and installed in the docker images.
To avoid a full dependency on franka_ros (only franka_description is required), these
packages are directly installed to /opt/ros in our docker images.
For the remaining CI workflow, we can stick with the standard process then.