RViz RS & Traj display: ensure robot_description is available onEnable
This is a (in my opinion) rather dirty workaround because RViz
enables Displays systematically *before* loading their properties.

So the patch loads the required property a second time *before* everything else.

Assuming this ever changes we can remove this code again.

A Display's "enabled" state is encoded in a boolean property the Display inherits from.
The way loading works (recursively running through the properties)
this special property is loaded first and triggers onEnable because
of the Property's changed() signal.

A second "Enabled" property exists in RViz's Displays but does not change behavior at all.