Update subframes test
Sometime the SubframesTests will fail. Planning on this particular problem
takes ~10s and the allotted planning time for move_group was set to 10s.
The planning time has been increased to 30s and the target pose offset was
increased from 1cm to 5cm. This should make the planning problem easier to
solve and therefore more likely to pass the test.