MGI test: synchronize async methods for testing purposes
In order to synchronize we subscribe to the monitored_planning_scene of
the move_group node and wait for geometry updates.
Because the test is the only thing interacting with the `move_group`,
it should be ok to wait for a single geometry update in each case.

Additionally a short sleep needs to be added after creating MGI & PSI
to allow the `move_group` node to subscribe. Otherwise the first test might fail
just because the message was dropped.