collision_detection_fcl: Report link_names in correct order (#2682)
* core/collision_detection_fcl: Fix link order in DistanceResult

FCL does not guarantee that fcl::distance(o1, o2) returns data in the
order o1,o2. So use the result to figure out link names and body types.

Furthermore, we do not compute normals in the signed distance case
anymore, unless enable_nearest_points is set. This mirrors the behavior
in the unsigned distance case.

* core/collision_detection_fcl: unit test for nearest_points

* collision_detection_fcl: warn user about bogus nearest_points on old FCL