Migrate PRBT ikfast plugin from moveit_resources
This circumnavigates problems with running tests in the moveit repository
that are based on this plugin. Before, this plugin was part of
moveit_resources, but introduced a compile time dependency on moveit_core.

Because a repository's tests should run without having to compile
downstream code (a sane requirement on buildfarm/prerelease tests),
we move the plugin here instead.