Fix an off-by-one error in servo_calcs.cpp
When both:

use_gazebo: true
command_out_type: trajectory_msgs/JointTrajectory

are set in the parameters a trajectory with redundant points is
published by servo_server. The 2nd point in this trajectory is always
empty (i.e. 0 timestamp and empty position/velocity/acceleration
arrays). This turns out to be because it is erroneously never filled in.
The timestamps need to be shifted relative to the point's index but all
of the points should still be filled in.