support arbitrary realtime-factors in trajectory visualization
The fact that trajectory visualization ignored the time parameterization by default
and that the only way to change this is the special value "REALTIME" always bothered me.

I changed the default to a realtime factor of 3x to get a nice preview of the trajectory
when using the "Plan&Execute" button.