use $(dirname) in MoveIt Setup Assistant (#2748)
Available since lunar. This simplifies the descriptions in moveit_resources
quite a bit because their package name does not match the include path
and we have to adapt the names everywhere after edits.

Note that two candidates for this replacement still remain -
they define the file path to the rviz config in an arg value.
Due to [a bug](https://github.com/ros/ros_comm/issues/1487)
these two can not be replaced at this moment.