Support timed mutex locking in PlanningSceneMonitor
This replaces the shared_mutex of the PSM with a shared_timed_mutex
that allows specifying a timeout for attempting to lock the mutex
for read and write using the LockedPlanningSceneRO and
LockedPlanningSceneRW classes. In case of a failed attempt, a
TryLockTimeoutException is thrown so that the calling thread is
unblocked in case of concurrency issues.
Additionally, a new scoped lock wrapper class TimedLock allows
tracking lock sources (=caller names), start time, waiting time and
the duration that the lock persists. An optional DiagnosticStatus
publisher can be enabled for debugging these lock events at runtime.

The scope-less locking functions (lockSceneRead(), etc...) are deprecated
since they are error-prone and don't work well with the diagnostics feature.