MPD: do not save/restore warehouse parameters
If we reload these values from the config, setting the ROS parameters is much less useful.
At least the *type* of warehouse_ros plugin (mongo or sqlite) cannot be selected
in the display, so you will probably need to meddle with the parameters anyway if you want
to connect to a different db.

search for parameters warehouse_host/port because they are usually set at the top level, but
you might want to set them differently for different move_groups.