Support new configuration scheme with multiple pipelines
Traditionally, PlanningPipeline parameters were defined in the private namespace of the move_group node.
Since MoveIt 1.1.2 multiple pipeline configs are supported in parallel, defined as ~planning_pipelines/*.
To support this new scheme as well (in a minimal fashion), we check for presence of the new scheme
and use the parameters of the default pipeline, specified via the parameter ~default_planning_pipeline.