TOTG: catch division by 0
This bug is already in the original implementation:

https://github.com/tobiaskunz/trajectories/blob/master/Path.cpp

In case the dot product between the two vectors is close to +/-1,
angle becomes +/-PI and cos/tan of 0.5 * PI in the lines below will
produce a division by 0.

This happens easily if a optimal trajectory is processed by TOTG, i.e.,
if a trajectory is processed by TOTG twice.