MPD: Avoid flickering of the progress bar
The progress bar shows the number of pending background jobs.
If there is only one job pending, the progress bar is shown and
immediately hidden as soon as the process is finished.
Thus, we shouldn't show the progress bar if there is only one job
and thus no actual progress to show.

Use the default size and color scheme.