planning_context_manager: rename protected methods
the trailing 1 and 2 and bind() usage made this harder to parse than it has to be.

Technically this breaks API (the methods are protected, not private).
But if you are that far into the code base to inherit here and rely on the names
of these simple methods, you probably agree that this improves readability.