Fix memory leak in mesh filter (#3371)
The mesh filter was allocating a new data pointer (for conversion from GL_FLOAT to GL_UNSIGNED_SHORT) on each call. This data pointer was to be cached, but the corresponding `data_size` variable was never updated, such that the cached pointer was never reused. As this kind of caching is not thread-safe and because a data conversion turns out to be not needed, the whole code was removed and replaced by a direct call of the filter on the raw float data.

Co-authored-by: Robert Haschke <rhaschke@techfak.uni-bielefeld.de>