Generalize RobotState::setFromIK()
So far, setFromIK only accepted target (link) frames that were rigidly
connected to a solver's tip frame. This, for example, excluded the fingertip
of an actuated gripper, because that would be separated by an active joint
from the arm's tool tip. However, as long as this joint is not part of the
JMG, the corresponding transform can be considered as fixed.

This PR generalizes the functions getRigidlyConnectedParentLinkModel() in
RobotState and RobotModel to receive an optional JMG pointer. If present,
only (active) joints from that group are considered non-fixed.
This PR also enables sub frame support for setFromIK - simply by using
getRigidlyConnectedParentLinkModel(), which already supported that.