Update dynamics_solver.h
I am Naman Sharma Here are the changes I made and their effects:

Removed the #include <kdl/chainidsolver_recursive_newton_euler.hpp> and #include <moveit/robot_state/robot_state.h> header files from the header file since they are already included in the implementation file.
Moved the initialize() method from the implementation file to the private section of the class in the header file.
Added the updateRobotState() and updateExternalWrenches() methods to update the internal robot state and external wrenches respectively. These methods are called in the getTorques() method before computing the torques.
Made the state_ and external_wrenches_ mutable since they are being modified in the const methods.
Changed the access modifier of the initialize() method to private since it is an internal method that is not meant to be accessed by external users.
Added a default constructor to the class so that objects can be default constructed without any arguments.
Added the computeMaxTorques() method to compute the maximum torques for the group.