Correct the inconsistent angle definition between IKFast solvers and the plugin template
Problem Description:

When determining the angle parameter to a IKFast solver of a type
IKP_Translation*AxisAngle4D/IKP_Translation*AxisAngle*Norm4D,
the current implementation takes the roll, pitch, and yaw value
from the pose frame's orientation directly in each respective
case, which does not match how angles are calculated in IKFast.

In the solver-generating script of IKFast (ikfast.py), these
angles are defined as (in the base link's frame):

1. the angle between the given axis (x/y/z) and the manipulator's
direction vector, or
2. when the corresponding normal vector (z/x/y) is set, the angle
between the given axis (x/y/z) and the projection of the
manipulator's direction vector onto the plane of which the normal
is the aforementioned one.

Therefore, for example, if the orientation of a pose frame is
described by (roll = PI, pitch = 0, yaw = 0), a direction vector
(0, 0, 1) in the manipulator's frame will be mapped to (0, 0, -1);
in that case, a TranslationXAxisAngle4D solver will view it as
PI/2 from the x-axis, while the current implementation only comes
up with 0 angle.

Solution:

This patch has implemented exactly the same way of angle
calculation as done in IKFast, based on the assumption of where
the manipulator's direction vector points to in its own frame.
Note that this assumption is identical to the ones made in other
cases, like IKP_Ray4D or IKP_TranslationDirection5D.